pyproject.toml
[project]
name = "policy-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.4.1",
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.7",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-groq>=1.1.1",
    "pymupdf>=1.26.7",
    "pypdf>=6.6.2",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.2",
]
